+++ src/validator.js
@@ -1,6 +1,5 @@
 define(function (require, exports, module) {
   var Core = require('./core'),
-      Widget = require('widget'),
       $ = require('$');
 
   var Validator = Core.extend({