+++ HISTORY.md
@@ -2,6 +2,29 @@
 
 ---
 
+
+## 0.9.7
+
+`tag:improved` Item 添加 `getMessage(theRule, isSuccess, options)` , 用于返回某规则的成功/错误提示信息.
+
+## 0.9.6
+
+`tag:fixed` [#46](https://github.com/aralejs/validator/issues/46)
+
+`tag:fixed` [#43](https://github.com/aralejs/validator/issues/43)  required 配置项添加函数支持, 默认为 false, 如果为 Function, 则根据 Function 返回的值来校验是否必要
+
+`tag:fixed` destroy 时, 调整 earse 和 hideMessage 的顺序, 因为在 hideMessage 中有可能会调用 this.query(element), 这样如果先删除了, 这边查询不到.
+
+## 0.9.5
+
+`tag:fixed` [#37](https://github.com/aralejs/validator/issues/37) errormessage 重写后无法替换自定义的 {{}}
+
+`tag:fixed` [#39](https://github.com/aralejs/validator/issues/39) DOM 有 disabled 属性的 item, 不做校验
+
+`tag:fixed` [#40](https://github.com/aralejs/validator/issues/40) 去除对 `:visible` 伪类的依赖.
+
+`tag:fixed` [#31](https://github.com/aralejs/validator/issues/31) addRule 方法回调函数 this 指向当前 Validtor 实例对象
+
 ## 0.9.4
 
 `tag:improved` 对 arale.widget 的依赖从 1.0.0 升级到 1.1.1
@@ -63,3 +86,4 @@
 `tag:improved` 对 widget 的依赖从 1.0.0 变为 1.0.2
 
 `tag:changed` addRule 的时候，同名规则可以覆盖了（0.8.8之前不能覆盖）
+